git-svn-id: https://svn.kuali.org/repos/foundation/branches/impex-2.0@19066 eb505253-de9c-4f3b-bf87-3aad84f29c0f

+++ torque-generator/src/main/java/org/kuali/common/impex/spring/BatchConfig.java
@@ -14,9 +14,9 @@ public class BatchConfig {
 	Environment env;
 
 	@Bean
-	public long impexBatchSize() {
+	public int impexBatchSize() {
 		String size = SpringUtils.getProperty(env, "impex.batch.size");
-		return FormatUtils.getBytes(size);
+		return (int) FormatUtils.getBytes(size);
 	}
 
 	@Bean
